# apps/backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/create": {
      "post": {
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Opération réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un nouveau projet.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects": {
      "get": {
        "operationId": "ProjectsController_getProject",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Opération réussie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les projets de l'utilisateur connecté.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{projectNumber}": {
      "get": {
        "operationId": "ProjectsController_getProjectByNumber",
        "parameters": [
          {
            "name": "projectNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opération réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un projet par son numéro.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/prompt": {
      "put": {
        "operationId": "ProjectsController_updateProjectPrompt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Opération réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mise à jour des instructions de prompt pour un projet.",
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Le Journal API",
    "description": "Documentation de l'API Le Journal",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {}
      },
      "ProjectUpdate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "ID"
          },
          "promptInstruction": {
            "type": "string",
            "description": "The instruction prompt for the project"
          },
          "canUpdatePrompt": {
            "type": "boolean",
            "description": "Whether the prompt can be updated"
          },
          "lastPromptUpdate": {
            "type": "string",
            "description": "The last prompt update date"
          }
        },
        "required": [
          "id",
          "promptInstruction",
          "canUpdatePrompt",
          "lastPromptUpdate"
        ]
      },
...
    }
  }
}